+++ tests/10_sealed_object_test.rs
@@ -19,7 +19,7 @@ use rand::SeedableRng;
 use rand_chacha::ChaChaRng;
 use serial_test::serial;
 use std::num::NonZeroUsize;
-use tss2_fapi_rs::{FapiContext, SealFlags};
+use tss2_fapi_rs::{FapiContext, SealData, SealFlags};
 
 const SEAL_TYPE_FLAGS: &[SealFlags] = &[SealFlags::NoDA];
 
@@ -27,11 +27,11 @@ const SEAL_TYPE_FLAGS: &[SealFlags] = &[SealFlags::NoDA];
 // Test cases
 // ==========================================================================
 
-/// Test the `create_seal()` function to create a new sealed object
+/// Test the `create_seal()` function to create a new sealed object containing random data.
 #[test]
 #[serial]
 #[named]
-fn test_create_seal() {
+fn test_create_seal_random() {
     let _configuration = TestConfiguration::with_finalizer(|| my_tpm_finalizer(PASSWORD));
 
     repeat_test!(|i| {
@@ -47,7 +47,40 @@ fn test_create_seal() {
         tpm_initialize!(context, PASSWORD, MyCallbacks::new(PASSWORD, None));
 
         // Create new seal, if not already created
-        match context.create_seal(key_path, Some(SEAL_TYPE_FLAGS), NonZeroUsize::new(32usize).unwrap(), None, None, None) {
+        match context.create_seal(key_path, Some(SEAL_TYPE_FLAGS), None, None, SealData::from_size(NonZeroUsize::new(32usize).unwrap())) {
+            Ok(_) => debug!("Seal created."),
+            Err(error) => panic!("Seal creation has failed: {:?}", error),
+        }
+    });
+}
+
+/// Test the `create_seal()` function to create a new sealed object containing application-defined data.
+#[test]
+#[serial]
+#[named]
+fn test_create_seal_data() {
+    let _configuration = TestConfiguration::with_finalizer(|| my_tpm_finalizer(PASSWORD));
+
+    repeat_test!(|i| {
+        let key_path = &format!("/HS/SRK/mySealObj{}", i);
+
+        // Initialize RNG
+        let mut rng = ChaChaRng::from_seed(create_seed(i));
+
+        // Create FAPI context
+        let mut context = match FapiContext::new() {
+            Ok(fpai_ctx) => fpai_ctx,
+            Err(error) => panic!("Failed to create context: {:?}", error),
+        };
+
+        // Initialize TPM, if not already initialized
+        tpm_initialize!(context, PASSWORD, MyCallbacks::new(PASSWORD, None));
+
+        // Generate plain-text
+        let sealed_data: [u8; 32usize] = generate_bytes(&mut rng);
+
+        // Create new seal, if not already created
+        match context.create_seal(key_path, Some(SEAL_TYPE_FLAGS), None, None, SealData::from_data(&sealed_data[..]).unwrap()) {
             Ok(_) => debug!("Seal created."),
             Err(error) => panic!("Seal creation has failed: {:?}", error),
         }
@@ -80,7 +113,7 @@ fn test_unseal() {
         let original_data: [u8; 128usize] = generate_bytes(&mut rng);
 
         // Create new seal, if not already created
-        match context.create_seal(key_path, Some(SEAL_TYPE_FLAGS), NonZeroUsize::new(original_data.len()).unwrap(), None, None, Some(&original_data[..])) {
+        match context.create_seal(key_path, Some(SEAL_TYPE_FLAGS), None, None, SealData::from_data(&original_data[..]).unwrap()) {
             Ok(_) => debug!("Seal created."),
             Err(error) => panic!("Seal creation has failed: {:?}", error),
         }